Feat: Make AI align 2 cells
If the attempt to find a possibility to win or counter failed, it tries to find a way to align an owned cell to another one. (Next to it or with a neutral cell in the middle)
Since there can be multiple possible lines to align the cell, a random line is picked. Then a random cell in this line is also picked. (Avoiding the non neutral cell)